backend/requirements.txt
# Core Framework
streamlit==1.31.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# AI & Data Processing
google-generativeai==0.3.2

# Data Handling
pandas==1.5.3
openpyxl==3.1.2
xlrd==2.0.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1
statsmodels==0.14.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==8.4.2
pytest-cov==7.0.0